- Added Docs - Moved Transmission Data classes into invoice header subpackage
