+++ src/main/java/com/github/siboxd/fatturapa/model/DatiAnagraficiCessionarioType.java
@@ -1,6 +1,6 @@
 package com.github.siboxd.fatturapa.model;
 
-import com.github.siboxd.fatturapa.model.invoiceheader.IdFiscaleType;
+import com.github.siboxd.fatturapa.model.invoiceheader.transmissiondata.IdFiscaleType;
 
 import org.simpleframework.xml.Element;
 import org.simpleframework.xml.Root;
