+++ src/main/java/com/github/siboxd/fatturapa/model/FatturaElettronica.java
@@ -1,6 +1,6 @@
 package com.github.siboxd.fatturapa.model;
 
-import com.github.siboxd.fatturapa.model.invoiceheader.FatturaElettronicaHeaderType;
+import com.github.siboxd.fatturapa.model.invoiceheader.transmissiondata.FatturaElettronicaHeaderType;
 
 import org.simpleframework.xml.Attribute;
 import org.simpleframework.xml.Element;
